re: MONDRIAN-1656.  Workaround for the zero-length arrays.
Simply check for a zero length array before accessing the element.  Added
a JUnit test for this issue that replicates the exception.  The MS Access
food mart database attached to this commit must be used.  All values for
the new column must be null.  I haven't tested it on any of the other
sample data sources, so the modified schema attached will not work with
them.